completed the template management page

+++ app/Http/Controllers/TemplateController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\HTTP\Response;
 use App\Models\Message;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
+
 
 class TemplateController extends Controller
 {
@@ -13,12 +16,26 @@ class TemplateController extends Controller
      */
     public function index()
     {
-        $messages = Message::all();
-        $messages->transform(function ($message) {
-            $message->content = strip_tags($message->content);
-            return $message;
-        });
-        return view('template.index', compact('messages'));
+        // $user = Auth::user();
+        // $messages = Message::all();
+        // $messages->transform(function ($message) {
+        //     $message->content = strip_tags($message->content);
+        //     return $message;
+        // });
+        // return view('template.index', compact('messages'));
+
+
+        $user = Auth::user();
+        if ($user) {
+            $messages = Message::where("writer", $user->email)
+                ->orWhere('editable', 0)->orderBy('id', 'desc')->get();
+            $messages->transform(function ($message) {
+                $message->content = strip_tags($message->content);
+                return $message;
+            });
+            return view('template.index', compact('messages'));
+        }
+        redirect("home");
     }
 
     /**
@@ -34,19 +51,58 @@ class TemplateController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title' => 'required',
-            'type' => 'required',
-            'trigger' => 'required',
-            'content' => 'required'
-        ]);
+        $user = Auth::user();
+        if ($user) {
+            $request->validate([
+                'title' => 'required',
+                'type' => 'required',
+                'trigger' => 'required',
+                'content' => 'required'
+            ]);
+            $request['writer'] = $user->email;
 
-        $save_data = Message::create($request->all());
+            $save_data = Message::create($request->all());
 
-        return response()->json([
-            'status' => 'success',
-            'data' => $save_data,
-        ], Response::HTTP_OK);
+            return response()->json([
+                'status' => 'success',
+                'data' => $save_data,
+            ], Response::HTTP_OK);
+        }
+        redirect("home");
+    }
+
+    /**
+     * copy a newly created resource in storage.
+     */
+    public function copy($id)
+    {
+        $user = Auth::user();
+        if ($user) {
+            $message = Message::find($id);
+            $new_message = [
+                'title' => $message->title,
+                'type' =>  $message->type,
+                'trigger' => $message->trigger,
+                'content' => $message->content,
+                'writer' => $user->email,
+            ];
+
+            $save_data = Message::create($new_message);
+            Log::info($message);
+
+            if ($save_data) {
+                return response()->json([
+                    'status' => 'success',
+                    'data' => $save_data,
+                ], Response::HTTP_OK);
+            }
+            return response()->json([
+                'status' => 'failed',
+                'data' => $save_data,
+                'message' => '操作が失敗しました。',
+            ], Response::HTTP_SERVICE_UNAVAILABLE);
+        }
+        redirect("home");
     }
 
     /**
@@ -84,18 +140,27 @@ class TemplateController extends Controller
     public function destroy(string $id)
     {
         $message = Message::find($id);
+        $email = Auth::user()->email;
         if ($message) {
-            if ($message->editable == 1) {
-                $message->delete();
+            if ($email == $message->writer) {
+                if ($message->editable == 1) {
+                    $message->delete();
+                    return response()->json([
+                        'status' => 'success',
+                        'message' => '削除操作が正常に行われました。'
+                    ], Response::HTTP_OK);
+                }
                 return response()->json([
-                    'status' => 'success',
-                    'message' => '削除操作が正常に行われました。'
-                ], Response::HTTP_OK);
+                    'status' => 'failed',
+                    'message' => 'このメッセージは削除できません。'
+                ], Response::HTTP_BAD_REQUEST);
+            } else {
+                return response()->json([
+                    'status' => 'failed',
+                    'message' => 'このメッセージを削除する権限がありません。'
+
+                ], Response::HTTP_BAD_REQUEST);
             }
-            return response()->json([
-                'status' => 'failed',
-                'message' => 'このメッセージは削除できません。'
-            ], Response::HTTP_BAD_REQUEST);
         }
         return response()->json([
             'status' => 'failed',
